refactor attack display
