+++ ElectronicObserver/Window/Tools/SortieRecordViewer/Sortie/Battle/Phase/AirBattleAttackViewModel.cs
@@ -7,7 +7,7 @@ using ElectronicObserverTypes.Attacks;
 
 namespace ElectronicObserver.Window.Tools.SortieRecordViewer.Sortie.Battle.Phase;
 
-public class AirBattleAttackViewModel
+public sealed class AirBattleAttackViewModel : AttackViewModelBase
 {
 	private int WaveIndex { get; }
 	public BattleIndex DefenderIndex { get; }
@@ -53,22 +53,5 @@ public class AirBattleAttackViewModel
 		AirAttack.Bombing => ConstantsRes.BombingAttack,
 		AirAttack.TorpedoBombing => ConstantsRes.TorpBombingAttack,
 		_ => ConstantsRes.Unknown,
-	};
-
-	private static string AttackDisplay(bool guardsFlagship, double damage, HitType hitType) => hitType switch
-	{
-		HitType.Hit => $"{HitDisplay(guardsFlagship, damage)} Dmg",
-		HitType.Critical => $"{HitDisplay(guardsFlagship, damage)} Critical!",
-		HitType.Miss => "Miss",
-		_ => "",
-	};
-
-	private static string HitDisplay(bool guardsFlagship, double damage)
-		=> $"{ProtectedDisplay(guardsFlagship)}{damage}";
-
-	private static string ProtectedDisplay(bool guardsFlagship) => guardsFlagship switch
-	{
-		true => $"<{BattleRes.Protected}> ",
-		_ => "",
 	};
 }
